Functional implementation of match list based indentation -- see issue #1
Still need to do some code cleanup.

Also, found that "else" gets confused when the ("else" ("if" "endif" "else"))
structure is in place. This way, it thinks that and if-else-endif block has 2
ending keywords, so it looks for the 3rd "if". If this is going to be fixed,
there has to be a privileged ending keyword that clears the unprivileged
keyword counts.